Made it so you can zoom in 8x in the editor. Changed default to 2x from 1x.
Default level zoom is 2x.
Previous editor max zoom was 1x.
Editor min zoom remains 0.16x.

+++ src/editor.hpp
@@ -187,7 +187,7 @@ public:
 	void autosave_level();
 	void zoomIn();
 	void zoomOut();
-	int zoom() const { return zoom_; }
+	float zoom() const { return zoom_; }
 
 	void undo_command();
 	void redo_command();
@@ -337,7 +337,7 @@ protected:
 	LevelPtr lvl_;
 
 	std::vector<LevelPtr> levels_;
-	int zoom_;
+	float zoom_;
 	int xpos_, ypos_;
 	int anchorx_, anchory_;
